pyproject.toml
[tool.poetry]
name = "account"
version = "0.1.0"
description = "signup/signin"
authors = ["Vitaliy Shishorin <moskrc@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.1"
django = "^3.1"
django-rest-framework = "^0.1.0"
django-environ = "^0.4.5"
django-cors-headers = "^3.5.0"
djangorestframework-simplejwt = "^4.4.0"

[tool.poetry.dev-dependencies]
mypy = "^0.782"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
